requirements.txt
Library:
streamlit
tensorflow
pillow
numpy